requirements.txt
accelerate==0.20.3
ftfy
gdown
matplotlib
numpy
opencv-python
pandas
Pillow
ruamel.yaml
scipy
scikit-learn
seaborn
tokenizers==0.13.3
tqdm
transformers==4.25.0
torch_fidelity
torch==2.0.1
wandb
NudeNet==3.0.8
fire
git+https://github.com/openai/CLIP.git
